Move unit tests into subfolder so we can add an integration test
