+++ composer.json
@@ -29,7 +29,7 @@
     },
     "autoload-dev": {
       "psr-4": {
-        "Scoutapm\\Tests\\": "tests/"
+        "Scoutapm\\UnitTests\\": "tests/Unit/"
       }
     },
     "config" : {
