+++ tests/Unit/Helper/BacktraceTest.php
@@ -1,5 +1,5 @@
 <?php
-namespace Scoutapm\Tests\Helper;
+namespace Scoutapm\UnitTests\Helper;
 
 use \PHPUnit\Framework\TestCase;
 use \Scoutapm\Helper\Backtrace;
