--- tests/bootstrap.php
@@ -1,4 +1,0 @@
-<?php
-namespace Scoutapm\Tests {
-    require __DIR__ . '/../vendor/autoload.php';
-}
